fix files/classes in wrong namespace
- move iterators from the generic alpaka namespace to `alpaak::onAcc`
- split `Iter.hpp` which contains many class definitions in seperate
  files
- move internal classes to the internal namespace
