+++ include/alpaka/alpaka.hpp
@@ -24,14 +24,15 @@
 #include "alpaka/math/Complex.hpp"
 #include "alpaka/math/constants.hpp"
 #include "alpaka/mem/BoundaryIter.hpp"
-#include "alpaka/mem/Iter.hpp"
 #include "alpaka/mem/LinearizedIdxGenerator.hpp"
 #include "alpaka/onAcc/Acc.hpp"
 #include "alpaka/onAcc/SimdAlgo.hpp"
+#include "alpaka/onAcc/WorkerGroup.hpp"
 #include "alpaka/onAcc/atomic.hpp"
 #include "alpaka/onAcc/globalMem.hpp"
 #include "alpaka/onAcc/interface.hpp"
 #include "alpaka/onAcc/memFence.hpp"
+#include "alpaka/onAcc/range.hpp"
 #include "alpaka/onAcc/tag.hpp"
 #include "alpaka/onHost/Device.hpp"
 #include "alpaka/onHost/DeviceSelector.hpp"
