+++ include/alpaka/onAcc/internal/SimdConcurrent.hpp
@@ -11,6 +11,7 @@
 #include "alpaka/core/common.hpp"
 #include "alpaka/mem/concepts.hpp"
 #include "alpaka/mem/concepts/IGeneratorOrMdSpan.hpp"
+#include "alpaka/onAcc/WorkerGroup.hpp"
 #include "alpaka/onAcc/interface.hpp"
 
 #include <cstdint>
