+++ include/alpaka/onAcc/internal/SimdTransformReduce.hpp
@@ -13,6 +13,7 @@
 #include "alpaka/mem/concepts.hpp"
 #include "alpaka/mem/concepts/IGeneratorOrMdSpan.hpp"
 #include "alpaka/onAcc/Acc.hpp"
+#include "alpaka/onAcc/WorkerGroup.hpp"
 #include "alpaka/onAcc/interface.hpp"
 
 #include <cstdint>
